+++ src/cli/mod.rs
@@ -1,3 +1,4 @@
+pub(crate) mod align_seqs;
 pub(crate) mod cluster_structs;
 pub(crate) mod contacts;
 pub(crate) mod dsasa;